adding newsletter form component
